LocalAPI integration (#12)
* Add localApi. Extract Innobus, Aidoo, and localapi to their custom files

* Use of python-airzone 0.8.1

* Ensure that I/O is made out of main thread

* Transform get_devices into an async method

* Fix async call to method.

Fix wrong use of the async wrapper

* Properly import local modules

* Fix property setter

* Fix wrong machine id

* Fixed new API for getting zones from a machine

* Temp.fix for unique id

* Fix machine modes

* Fix machine operation mode

* Add fan speeds

* Add humidity, fix update

* Use coordinator for localapi

* update python-airzone version

* Fix super call

* remove coordinator

* update python airzone version

* Change method name

* Update python-airzone version

* Update python-airzone version

* Zones now retrieve data indepentdely

* Adapt innobus to use changes from localapi

* Update README
