+++ custom_components/airzone/const.py
@@ -2,6 +2,7 @@ DOMAIN = "airzone"
 DEFAULT_DEVICE_ID = 1
 DEFAULT_DEVICE_CLASS = 'innobus'
 DEFAULT_SPEED_AS_PER = False
+SYSTEM_TYPES = ["innobus", "aido", "localapi"]
 
 from homeassistant.components.climate.const import (
     FAN_AUTO,
@@ -67,6 +68,40 @@ PRESET_AIR_MODE = 'AIRE'
 PRESET_FLOOR_MODE = 'FLOOR'
 MACHINE_PRESET_MODES = [PRESET_AIR_MODE, PRESET_FLOOR_MODE]
 MACHINE_SUPPORT_FLAGS = SUPPORT_AUX_HEAT | SUPPORT_PRESET_MODE
+
+# LocalAPI Modes
+
+LOCALAPI_ZONE_HVAC_MODES = [HVAC_MODE_HEAT_COOL, HVAC_MODE_OFF]
+LOCALAPI_ZONE_SUPPORT_FLAGS =  SUPPORT_TARGET_TEMPERATURE
+
+LOCALAPI_MACHINE_SUPPORT_FLAGS =  SUPPORT_FAN_MODE
+
+LOCALAPI_MACHINE_HVAC_MODES = [HVAC_MODE_OFF, 
+                            HVAC_MODE_COOL, 
+                            HVAC_MODE_HEAT, 
+                            HVAC_MODE_FAN_ONLY, 
+                            HVAC_MODE_DRY,
+                            HVAC_MODE_AUTO] 
+
+LOCALAPI_HVAC_MODE_MAP = {
+    HVAC_MODE_OFF:'STOP',
+    HVAC_MODE_COOL:'COOLING',    
+    HVAC_MODE_HEAT:'HEATING',
+    HVAC_MODE_FAN_ONLY:'FAN',
+    HVAC_MODE_DRY:'DRY',
+    HVAC_MODE_AUTO:'AUTO',
+}
+
+LOCALAPI_MODE_TO_HVAC_MAP = {
+    'STOP': HVAC_MODE_OFF,
+    'COOLING': HVAC_MODE_COOL,
+    'AUTO': HVAC_MODE_AUTO,
+    'HEATING': HVAC_MODE_HEAT,
+    'FAN': HVAC_MODE_FAN_ONLY,
+    'DRY': HVAC_MODE_DRY
+}
+
+
 
 #AIDO extra modes
