+++ custom_components/airzone/translations/en.json
@@ -1,7 +1,7 @@
 {
     "config": {
         "error": {
-            "connection": "Cannot connect with airzone TCP modbus"
+            "connection": "Cannot connect with airzone machine"
           },
         "step": {
             "user": {
